refactor : courseId와 courseType 반영
